# openapi.json
{
  "components": {
    "schemas": {
      "AckIn": {
        "properties": {
          "ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "AckIn",
        "type": "object"
      },
      "AckOut": {
        "properties": {
          "acked": {
            "title": "Acked",
            "type": "integer"
          }
        },
        "required": [
          "acked"
        ],
        "title": "AckOut",
        "type": "object"
      },
...
      "ErrorOut": {
        "description": "Ошибка со строковым кодом.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetroRunDetail": {
        "properties": {
          "buffs": {
            "items": {},
            "title": "Buffs",
            "type": "array"
          },
          "duration_s": {
            "title": "Duration S",
            "type": "number"
          },
          "events": {
            "items": {},
            "title": "Events",
            "type": "array"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "grid": {
            "additionalProperties": true,
            "title": "Grid",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "path": {
            "items": {},
            "title": "Path",
            "type": "array"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "scenario_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Run Id"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step S"
          },
          "steps": {
            "title": "Steps",
            "type": "integer"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "vitals": {
            "items": {},
            "title": "Vitals",
            "type": "array"
          }
        },
        "required": [
          "id",
          "scenario_run_id",
          "started_at",
          "finished_at",
          "status",
          "outcome",
          "steps",
          "duration_s",
          "step_s",
          "buffs",
          "result",
          "summary",
          "grid",
          "path",
          "events",
          "vitals"
        ],
        "title": "MetroRunDetail",
        "type": "object"
      },
...
      "MetroRunSummary": {
        "properties": {
          "buffs": {
            "items": {},
            "title": "Buffs",
            "type": "array"
          },
          "duration_s": {
            "title": "Duration S",
            "type": "number"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "scenario_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Run Id"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step S"
          },
          "steps": {
            "title": "Steps",
            "type": "integer"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "id",
          "scenario_run_id",
          "started_at",
          "finished_at",
          "status",
          "outcome",
          "steps",
          "duration_s",
          "step_s",
          "buffs",
          "result",
          "summary"
        ],
        "title": "MetroRunSummary",
        "type": "object"
      },
      "MetroRunsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MetroRunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before"
          }
        },
        "required": [
          "items",
          "next_before"
        ],
        "title": "MetroRunsPage",
        "type": "object"
      },
      "NotificationOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "read": {
            "title": "Read",
            "type": "boolean"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "level",
          "code",
          "text",
          "read"
        ],
        "title": "NotificationOut",
        "type": "object"
      },
      "NotificationsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before"
          },
          "unread": {
            "title": "Unread",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "unread",
          "next_before"
        ],
        "title": "NotificationsPage",
        "type": "object"
      },
...
      "ReadIn": {
        "properties": {
          "up_to_id": {
            "minimum": 1.0,
            "title": "Up To Id",
            "type": "integer"
          }
        },
        "required": [
          "up_to_id"
        ],
        "title": "ReadIn",
        "type": "object"
      },
      "ReadOut": {
        "properties": {
          "read": {
            "title": "Read",
            "type": "integer"
          }
        },
        "required": [
          "read"
        ],
        "title": "ReadOut",
        "type": "object"
      },
...
      "UnrecognizedOut": {
        "properties": {
          "acked": {
            "title": "Acked",
            "type": "boolean"
          },
          "chat_id": {
            "title": "Chat Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "first_line": {
            "title": "First Line",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "message_id": {
            "title": "Message Id",
            "type": "integer"
          },
          "msg_id": {
            "title": "Msg Id",
            "type": "integer"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "id",
          "message_id",
          "chat_id",
          "msg_id",
          "first_line",
          "text",
          "created_at",
          "acked"
        ],
        "title": "UnrecognizedOut",
        "type": "object"
      },
      "UnrecognizedPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UnrecognizedOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before"
          }
        },
        "required": [
          "items",
          "next_before"
        ],
        "title": "UnrecognizedPage",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "pyrobot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/metro/runs": {
      "get": {
        "operationId": "metro_runs_api_v1_metro_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetroRunsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not authenticated"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Metro Runs",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/metro/runs/{run_id}": {
      "get": {
        "operationId": "metro_run_api_v1_metro_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetroRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "metro run not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Metro Run",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/notifications": {
      "get": {
        "operationId": "notifications_api_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "unread",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "info",
                    "warn",
                    "error"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not authenticated"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Notifications",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/notifications/read": {
      "post": {
        "operationId": "read_notifications_api_v1_notifications_read_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "csrf token mismatch"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Notifications",
        "tags": [
          "reference"
        ]
      }
    },
...
    "/api/v1/unrecognized": {
      "get": {
        "operationId": "unrecognized_api_v1_unrecognized_get",
        "parameters": [
          {
            "in": "query",
            "name": "acked",
            "required": false,
            "schema": {
              "default": "false",
              "enum": [
                "false",
                "true",
                "all"
              ],
              "title": "Acked",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnrecognizedPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not authenticated"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unrecognized",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/unrecognized/ack": {
      "post": {
        "operationId": "ack_unrecognized_api_v1_unrecognized_ack_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "csrf token mismatch"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ack Unrecognized",
        "tags": [
          "reference"
        ]
      }
    },
...
  }
}